Abstract out & improve error handling.
Errors from API calls are now handled in a bottleneck that passes along
things like network errors untouched (the err parameter in request()
callbacks) and is aware of how Recurly reports errors.

To do this we pass along the response from each network request to
handleRecurlyError() along with a list of http status codes we expect to
see from the call. Any errors are bundled up in a single error object.
If you get an error from a recurly call, it's now wrapped in a
RecurlyError object, with the original response from them in the 'error'
field.