fix: invalid_state error on Invoice#all (#59)
When calling Invoice.all(function), i was seeing this error:

```
state_invalid
Invalid state (must be one of open, processing, collected, failed, past_due)
```

This is because the state query param was defaulting to `active` which
is not an invoice state (though it is a Subscription state). I think it
should default to not passing a state filter by default (which basically
means all invoices).