test: fix the broken invoice test (#60)
This fixes the integration test that was failing on every run. I fixed
it by having the setup callback create an open invoice for the test to
find. It creates an account, then an adjustment, then an invoice.

I also noticed that the Account model file was requiring the account as
`Adjustment`.

I also added some properties to the adjustment that were
missing.